+++ src/main/java/com/devpeople/bapsim/global/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.devpeople.bapsim.global.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Integer id) {
+    public UserNotFoundException(Long id) {
         super("사용자를 찾을 수 없습니다. id=" + id);
     }
 }
